services/conversion-engine/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
Pillow==10.2.0
pdf2image==1.17.0
